Configure restart policy: 'unless-stopped' for containers in docker-compose.yml
This should autostart the containers -- as long as they have previously been created and started, typically by using the 'docker-compose up' command whenever the docker daemon itself is started.

It will also restart the containers if they exit - and that's potentially less ideal, depending on the reason for the container exit.